pyproject.toml
[tool.poetry]
name = "saprot"
version = "0.4.7"
description = ""
authors = ["Zhikai Li (modefied by Tianli Tao) <luckydogqaq@163.com>"]
readme = "README.md"


[tool.poetry.dependencies]
torch = "2.0.1"
torchvision = "0.15.2"
torchaudio = "2.0.2"
torchmetrics = "0.9.3"
pytorch-lightning = "2.1.3"
transformers = "4.28.0"
scikit-learn = "1.4.0"
tabulate = "0.9.0"
biopython = "1.83"
easydict = "1.13"
lmdb = "1.4.1"
peft = "0.10.0"
bio = "1.6.2"
wandb = "*"
scipy = "*"
pandas = "*"
fair-esm = "*"
matplotlib = "*"
multiprocess = "*"
loguru = "*"
colorama = "*"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
